Keep replay preservation within the chat size budget
The provider-history filter is a cohesive internal module, so extracting it keeps message preparation below its enforced complexity ceiling without changing behavior. The ratchet now locks in the smaller boundary.

Constraint: CI rejects src/chat/message-prep.ts above its line ceiling.

Rejected: Raise the line ceiling | hides structural growth and weakens an existing maintainability guard.

Confidence: high

Scope-risk: narrow

Directive: Keep provider-tool history filtering in provider-tool-history.ts rather than regrowing message-prep.ts.

Tested: Chat ratchet; 96 focused tests; Deno check; Deno lint; Deno fmt; API reference generation/check.

Not-tested: Full pre-push suite repeats during push.
